implement auth check in restaurants features

+++ routes/restaurants.js
@@ -8,10 +8,12 @@ const Restaurant = db.Restaurant
 router.get('/', (req, res, next) => {
   // console.log('session:', req.session)
   // console.log(req.user)
+  const userId = req.user.id
   const page = parseInt(req.query.page) || 1
   const limit = 9
 
   return Restaurant.findAndCountAll({
+    where: { userId },
     offset: (page - 1) * limit,
     limit,
     // order: sortCase,
@@ -41,12 +43,14 @@ router.get('/', (req, res, next) => {
 // 顯示 restaurant 清單頁(搜尋)
 router.get('/search', (req, res, next) => {
   const keywords = req.query.keyword?.trim()
+  const userId = req.user.id
   // console.log('keywords:', keywords)
   if (!keywords) {
     return res.redirect('/restaurants')
   }
 
   Restaurant.findAll({
+    where: { userId },
     raw: true
   })
     .then((restaurants) => {
@@ -64,9 +68,10 @@ router.get('/search', (req, res, next) => {
 
 // 顯示 restaurant 清單頁(分類)
 router.get('/', (req, res, next) => {
+  const userId = req.user.id
   const sort = req.query.sort || 'A'
 
-  const sortCase = function sortCase (sort) {
+  const sortCase = function sortCase(sort) {
     switch (sort) {
       case 'A':
         return [['name_en', 'ASC']]
@@ -84,6 +89,7 @@ router.get('/', (req, res, next) => {
   // console.log(sortCase(sort))
 
   return Restaurant.findAll({
+    where: { userId },
     order: sortCase(sort),
     raw: true
   })
@@ -102,7 +108,10 @@ router.get('/', (req, res, next) => {
 
 // 新增 restaurant 頁
 router.get('/new', (req, res, next) => {
+  const userId = req.user.id
+
   return Restaurant.findAll({
+    where: { userId },
     raw: true
   })
     .then((restaurant) => res.render('new', { restaurant }))
@@ -116,7 +125,11 @@ router.get('/new', (req, res, next) => {
 router.post('/', (req, res, next) => {
   // 從 req.body 中獲取表單資料
   const formData = req.body
-  // res.json(req.body)
+  const userId = req.user.id
+  // 將 userId 添加到 formData 中
+  formData.userId = userId
+  // console.log(formData)
+
   return Restaurant.create(formData)
     .then(() => {
       req.flash('success', '新增成功!')
@@ -132,10 +145,23 @@ router.post('/', (req, res, next) => {
 router.get('/:id', (req, res, next) => {
   const id = req.params.id
   // console.log('id:', req.params)
+
   return Restaurant.findByPk(id, {
     raw: true
   })
-    .then((restaurant) => res.render('detail', { restaurant }))
+    .then((restaurant) => {
+      if (!restaurant) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/restaurants')
+      }
+
+      if (restaurant.userId !== userId) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/restaurants')
+      }
+
+      res.render('detail', { restaurant })
+    })
     .catch((error) => {
       error.errorMessage = '資料取得失敗:('
       next(error)
@@ -145,10 +171,21 @@ router.get('/:id', (req, res, next) => {
 // 更新 restaurant 頁
 router.get('/:id/edit', (req, res, next) => {
   const id = req.params.id
+
   return Restaurant.findByPk(id, {
     raw: true
   })
     .then((restaurant) => {
+      if (!restaurant) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/restaurants')
+      }
+
+      if (restaurant.userId !== userId) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/restaurants')
+      }
+
       const category = restaurant.category
       // Handlebars.registerHelper 是 Handlebars.js 模板引擎中的一個方法
       // 用於註冊自定義的 Handlebars 輔助函數（helper functions）
@@ -156,7 +193,7 @@ router.get('/:id/edit', (req, res, next) => {
         // console.log(value)
         return value === category
       }
-      return res.render('edit', { restaurant, selectedCategory })
+      res.render('edit', { restaurant, selectedCategory })
     })
     .catch((error) => {
       error.errorMessage = '資料取得失敗:('
@@ -168,10 +205,27 @@ router.get('/:id/edit', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
   const formData = req.body
   const id = req.params.id
-  return Restaurant.update(formData, { where: { id } })
-    .then(() => {
-      req.flash('success', '更新成功!')
-      return res.redirect(`/restaurants/${id}`)
+  const userId = req.user.id
+
+  return Restaurant.findByPk(id, {
+    raw: true
+  })
+    .then((restaurant) => {
+      if (!restaurant) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/restaurants')
+      }
+
+      if (restaurant.userId !== userId) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/restaurants')
+      }
+
+      return Restaurant.update(formData, { where: { id } })
+        .then(() => {
+          req.flash('success', '更新成功!')
+          return res.redirect(`/restaurants/${id}`)
+        })
     })
     .catch((error) => {
       error.errorMessage = '更新失敗:('
@@ -182,10 +236,27 @@ router.put('/:id', (req, res, next) => {
 // 刪除 restaurant
 router.delete('/:id', (req, res, next) => {
   const id = req.params.id
-  return Restaurant.destroy({ where: { id } })
-    .then(() => {
-      req.flash('success', '刪除成功!')
-      return res.redirect('/restaurants')
+  const userId = req.user.id
+
+  return Restaurant.findByPk(id, {
+    raw: true
+  })
+    .then((restaurant) => {
+      if (!restaurant) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/restaurants')
+      }
+
+      if (restaurant.userId !== userId) {
+        req.flash('error', '找不到資料')
+        return res.redirect('/restaurants')
+      }
+
+      return Restaurant.destroy({ where: { id } })
+        .then(() => {
+          req.flash('success', '刪除成功!')
+          return res.redirect('/restaurants')
+        })
     })
     .catch((error) => {
       error.errorMessage = '刪除失敗:('
